kindle transcription/requirements.txt
# Core OCR
pytesseract==0.3.10
Pillow==10.4.0
pdf2image==1.17.0
opencv-python==4.10.0.84

# Google APIs
google-api-python-client==2.147.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-auth==2.35.0

# Utilities
python-dotenv==1.0.1
tqdm==4.66.5
loguru==0.7.2
tenacity==9.0.0

# Testing
pytest==8.3.3
pytest-mock==3.14.0
